+++ yaffs_checkptrw.c
@@ -87,7 +87,7 @@ static int yaffs2_checkpt_space_ok(struct yaffs_dev *dev)
 
 static int yaffs_checkpt_erase(struct yaffs_dev *dev)
 {
-	int i;
+	u32 i;
 
 	if (!dev->drv.drv_erase_fn)
 		return 0;
@@ -126,7 +126,7 @@ static int yaffs_checkpt_erase(struct yaffs_dev *dev)
 
 static void yaffs2_checkpt_find_erased_block(struct yaffs_dev *dev)
 {
-	int i;
+	u32 i;
 	int blocks_avail = dev->n_erased_blocks - dev->param.n_reserved_blocks;
 
 	yaffs_trace(YAFFS_TRACE_CHECKPOINT,
@@ -135,7 +135,7 @@ static void yaffs2_checkpt_find_erased_block(struct yaffs_dev *dev)
 		blocks_avail, dev->checkpt_next_block);
 
 	if (dev->checkpt_next_block >= 0 &&
-	    dev->checkpt_next_block <= dev->internal_end_block &&
+	    dev->checkpt_next_block <= (int)dev->internal_end_block &&
 	    blocks_avail > 0) {
 
 		for (i = dev->checkpt_next_block; i <= dev->internal_end_block;
@@ -160,7 +160,7 @@ static void yaffs2_checkpt_find_erased_block(struct yaffs_dev *dev)
 
 static void yaffs2_checkpt_find_block(struct yaffs_dev *dev)
 {
-	int i;
+	u32 i;
 	struct yaffs_ext_tags tags;
 
 	yaffs_trace(YAFFS_TRACE_CHECKPOINT,
@@ -210,7 +210,7 @@ static void yaffs2_checkpt_find_block(struct yaffs_dev *dev)
 
 int yaffs2_checkpt_open(struct yaffs_dev *dev, int writing)
 {
-	int i;
+	u32 i;
 
 	dev->checkpt_open_write = writing;
 
@@ -319,7 +319,7 @@ static int yaffs2_checkpt_flush_buffer(struct yaffs_dev *dev)
 				       dev->checkpt_buffer, &tags);
 	dev->checkpt_page_seq++;
 	dev->checkpt_cur_chunk++;
-	if (dev->checkpt_cur_chunk >= dev->param.chunks_per_block) {
+	if (dev->checkpt_cur_chunk >= (int)dev->param.chunks_per_block) {
 		dev->checkpt_cur_chunk = 0;
 		dev->checkpt_cur_block = -1;
 	}
@@ -354,7 +354,7 @@ int yaffs2_checkpt_wr(struct yaffs_dev *dev, const void *data, int n_bytes)
 		dev->checkpt_byte_count++;
 
 		if (dev->checkpt_byte_offs < 0 ||
-		    dev->checkpt_byte_offs >= dev->data_bytes_per_chunk)
+		    dev->checkpt_byte_offs >= (int)dev->data_bytes_per_chunk)
 			ok = yaffs2_checkpt_flush_buffer(dev);
 	}
 
@@ -378,7 +378,7 @@ int yaffs2_checkpt_rd(struct yaffs_dev *dev, void *data, int n_bytes)
 	while (i < n_bytes) {
 
 		if (dev->checkpt_byte_offs < 0 ||
-		    dev->checkpt_byte_offs >= dev->data_bytes_per_chunk) {
+		    dev->checkpt_byte_offs >= (int)dev->data_bytes_per_chunk) {
 
 			if (dev->checkpt_cur_block < 0) {
 				yaffs2_checkpt_find_block(dev);
@@ -403,7 +403,7 @@ int yaffs2_checkpt_rd(struct yaffs_dev *dev, void *data, int n_bytes)
 						&tags);
 
 			/* Bail out if the chunk is corrupted. */
-			if (tags.chunk_id != (dev->checkpt_page_seq + 1) ||
+			if (tags.chunk_id != (u32)(dev->checkpt_page_seq + 1) ||
 			    tags.ecc_result > YAFFS_ECC_RESULT_FIXED ||
 			    tags.seq_number != YAFFS_SEQUENCE_CHECKPOINT_DATA)
 				break;
@@ -416,7 +416,7 @@ int yaffs2_checkpt_rd(struct yaffs_dev *dev, void *data, int n_bytes)
 			dev->checkpt_cur_chunk++;
 
 			if (dev->checkpt_cur_chunk >=
-					dev->param.chunks_per_block)
+					(int)dev->param.chunks_per_block)
 				dev->checkpt_cur_block = -1;
 
 		}
@@ -435,7 +435,7 @@ int yaffs2_checkpt_rd(struct yaffs_dev *dev, void *data, int n_bytes)
 
 int yaffs_checkpt_close(struct yaffs_dev *dev)
 {
-	int i;
+	u32 i;
 
 	if (dev->checkpt_open_write) {
 		if (dev->checkpt_byte_offs !=
@@ -448,8 +448,8 @@ int yaffs_checkpt_close(struct yaffs_dev *dev)
 			int blk = dev->checkpt_block_list[i];
 			struct yaffs_block_info *bi = NULL;
 
-			if (dev->internal_start_block <= blk &&
-			    blk <= dev->internal_end_block)
+			if ((int)dev->internal_start_block <= blk &&
+			    blk <= (int)dev->internal_end_block)
 				bi = yaffs_get_block_info(dev, blk);
 			if (bi && bi->block_state == YAFFS_BLOCK_STATE_EMPTY)
 				bi->block_state = YAFFS_BLOCK_STATE_CHECKPOINT;
