+++ yaffs_getblockinfo.h
@@ -23,7 +23,8 @@
 static inline struct yaffs_block_info *yaffs_get_block_info(struct yaffs_dev
 							      *dev, int blk)
 {
-	if (blk < dev->internal_start_block || blk > dev->internal_end_block) {
+	if (blk < (int)dev->internal_start_block ||
+	    blk > (int)dev->internal_end_block) {
 		yaffs_trace(YAFFS_TRACE_ERROR,
 			"**>> yaffs: get_block_info block %d is not valid",
 			blk);
